+++ packages/pois/plugin.js
@@ -15,7 +15,7 @@ K.Plugin({
 			"pois": {
 				"limit": 30,
 				"caching": true,	//cache response by OSM overpass
-				"maxDistance": 600,
+				"maxDistance": 500,
 				"importPois": true,	//enable importing of pois like place
 				"typesByTags": {
 					"amenity=drinking_water": "water",