+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 ## Next version
 - new method K.Util.setPath set sub property of object by path string
 - new method K.Util.json2html transform json or literal object into a nested ul list
-- plugin pois new method K.Pois.poisTracks
+- plugin pois new async method K.Place.loadPoisTracks()
 - plugin Openrouteservice extend K.Pois.poisTracks
 - plugin Openrouteservice K.settings.public.openrouteservice.poisRoutes
+- plugin Openrouteservice K.settings.openrouteservice.caching
 - plugin admin new settings adminsAutoFriendship
 - plugin geoinfo new method K.Geoinfo.getIpInfo
 - plugin admin new method for debug data in database, sanitizePlacesField